rename viewer to node_factory, more explicit name
