+++ niviz/node_factory.py
@@ -2,7 +2,7 @@
 Provides interface to join objects generated by ConfigSpec
 to Nipype ReportCapableInterface
 
-Provides the niviz.view_adapter.factory object which can be used register and
+Provides the niviz.node_factory.factory object which can be used register and
 create interfaces
 """
 
@@ -46,7 +46,6 @@ class ArgInputSpec:
     name: str
     method: str
     interface_args: dict
-    bids_output: Path = None
 
     def __post_init__(self, out_path, bids_entities):
         '''
@@ -56,9 +55,13 @@ class ArgInputSpec:
             KeyError if BIDS entities required for output_path are missing
         '''
 
-        self.bids_output = Path(
-            Template(out_path).substitute({x[0]: x[1]
-                                           for x in bids_entities}))
+        self.interface_args.update({
+            'out_report':
+            Path(
+                Template(out_path).substitute(
+                    {x[0]: f"{x[0]}-{x[1]}"
+                     for x in bids_entities}))
+        })
 
 
 class RPTFactory(object):
@@ -96,14 +99,12 @@ class RPTFactory(object):
                 f"View method {spec.method} has not been registered "
                 "with RPTFactory. If using a custom ReportCapableInterface"
                 " register to RPTFactory using\n"
-                "from view_adapter import register_interface\n"
+                "from niviz.node_factory import register_interface\n"
                 f"register_interface(ReportCapableInterface, {spec.method})")
             raise
 
         # Create and configure node args
-        interface = interface_class(**spec.argmap)
-        interface.outputs.out_report = spec.bids_output
-        return interface
+        return interface_class(generate_report=True, **spec.interface_args)
 
     def register_interface(self,
                            rpt_interface: reporting.ReportCapableInterface,
@@ -140,9 +141,8 @@ class RPTFactory(object):
         return self._interfaces
 
 
-
-
 factory = RPTFactory()
+
 
 def register_interface(rpt_interface: reporting.ReportCapableInterface,
                        method: str) -> None:
@@ -161,26 +161,26 @@ def register_interface(rpt_interface: reporting.ReportCapableInterface,
     '''
     factory.register_interface(rpt_interface, method)
 
-# Avoid circular import problem
-def initialize_defaults():
-    import niviz.interfaces.views
-    niviz.interfaces.views._run_imports()
 
-
-def get_interfaces_from_specs(
-        arg_specs: list[ArgInputSpec]) -> reporting.ReportCapableInterface:
+def get_interface(spec: ArgInputSpec) -> reporting.ReportCapableInterface:
     '''
     Generate Nipype ReportCapableInterfaces from a list of
     ArgInputSpecs
 
     Args:
-        arg_specs:  Iterable of ArgInputSpecs containing instructions for
+        spec:  ArgInputSpec containing instructions for
         building visualization interfaces
 
     Returns:
         Iterable of Nipype ReportCapableInterfaces
     '''
-    return [factory.get_interface(a) for a in arg_specs]
+    return factory.get_interface(spec)
+
+
+# Avoid circular import problem
+def initialize_defaults():
+    import niviz.interfaces.views
+    niviz.interfaces.views._run_imports()
 
 
 initialize_defaults()
